+++ inst/project_templates/Rprofile_template.R
@@ -1,0 +1,4 @@
+
+source(here::here(
+  list.files(here::here(), "_dependencies.R$")
+))