+++ NEWS.md
@@ -105,3 +105,4 @@
 -   Added **ody_rc_timetravel** to load previous imports.
 -   Now, tokens are stored in .Renviron, so you only have to provide it when starting a new project or after regenerating your token.
 -   Added **ody_extract_km** to get the median and confidence intervals of survfit objects.
+-   Added **ody_proj_init** to set the basicc structure of a non-RedCap project.